requirements.txt
torch>=2.2.0
transformers>=4.40.0
datasets>=2.19.0
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.26.0
wandb>=0.17.0
